refactor: improve code style
