--- src/components/NavigationButton/NavigationButtonTypes.ts
@@ -1,8 +1,0 @@
-import { TabbarItemProps } from "@vkontakte/vkui/dist/components/TabbarItem/TabbarItem";
-import { HTMLAttributes, ReactNode } from "react";
-
-interface NavigationButtonTypes extends TabbarItemProps, HTMLAttributes<HTMLElement> {
-  children: ReactNode;
-}
-
-export default NavigationButtonTypes;
