+++ src/index.ts
@@ -1,3 +1,2 @@
-export { default as NavigationButton } from "./components/NavigationButton/NavigationButton";
-export { default as NavigationProvider } from "./components/NavigationProvider/NavigationProvider";
-export { default as NavigationButtonType } from "./components/NavigationButton/NavigationButtonTypes";
+export * from './types';
+export * from './components';
